# Conflicts:
#	app/build.gradle
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/MainActivity.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/calendar/CalendarCommunication.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/calendar/CalendarFragment.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/community/groups/InfoGroupCommunication.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/community/groups/PostsFragment.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/infopet/InfoPetMealsFragment.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/infopet/InfoPetMedicationFragment.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/chart/statisticdata/DailyKilocaloriesData.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/chart/statisticdata/ExerciseFrequencyData.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/chart/statisticdata/StatisticData.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/chart/statisticdata/WashFrequencyData.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/chart/statisticdata/WeeklyExerciseData.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/chart/statisticdata/WeeklyKilocaloriesData.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/chart/statisticdata/WeightData.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/circularentry/forums/ForumsComponentView.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/circularentry/subscriber/SubscriberComponentView.java
#	app/src/main/java/org/pesmypetcare/mypetcare/activities/views/circularentry/subscriber/SubscribersView.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/community/TrAddNewForum.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/community/TrAddNewPost.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/community/TrCreateNewGroup.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/community/TrDeletePost.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/pethealth/TrAddNewWashFrequency.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/pethealth/TrAddNewWeight.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/pethealth/TrDeleteWashFrequency.java
#	app/src/main/java/org/pesmypetcare/mypetcare/controllers/pethealth/TrDeleteWeight.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/community/forums/Forum.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/community/groups/Group.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/community/posts/Post.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/pets/Event.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/pets/Meals.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/pets/Medication.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/pets/PeriodEvent.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/pets/Pet.java
#	app/src/main/java/org/pesmypetcare/mypetcare/features/pets/PetHealthInfo.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/CommunityAdapter.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/CommunityService.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/MealManagerAdapter.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/MedicationManagerAdapter.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/PetManagerAdapter.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/PetManagerService.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/ServiceLocator.java
#	app/src/main/java/org/pesmypetcare/mypetcare/services/UserManagerAdapter.java
#	app/src/main/res/values-ca/strings.xml
#	app/src/main/res/values-es-rES/strings.xml
#	app/src/main/res/values/strings.xml
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrAddNewForum.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrAddNewPost.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrAddNewWashFrequency.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrAddNewWeight.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrAddSubscription.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrCreateNewGroup.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeleteForum.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeleteMeal.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeleteMedication.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeletePeriodicNotification.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeletePersonalEvent.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeletePet.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeletePost.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeleteSubscription.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeleteWashFrequency.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrDeleteWeight.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrNewPeriodicNotification.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrNewPersonalEvent.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrNewPetMeal.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrNewPetMedication.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrRegisterNewPet.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrUpdateMedication.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrUpdatePet.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrUpdatePetImage.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrUpdatePetMeal.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/TestTrUpdatePost.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/community/TestTrLikePost.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/community/TestTrReportPost.java
#	app/src/test/java/org/pesmypetcare/mypetcare/controllers/community/TestTrUnlikePost.java
#	app/src/test/java/org/pesmypetcare/mypetcare/features/pets/TestPet.java
#	app/src/test/java/org/pesmypetcare/mypetcare/services/StubCommunityService.java
#	app/src/test/java/org/pesmypetcare/mypetcare/services/StubMealManagerService.java
#	app/src/test/java/org/pesmypetcare/mypetcare/services/StubMedicationService.java
#	app/src/test/java/org/pesmypetcare/mypetcare/services/StubPetManagerService.java

+++ app/src/main/java/org/pesmypetcare/mypetcare/controllers/meals/TrDeleteMeal.java
@@ -2,17 +2,23 @@ package org.pesmypetcare.mypetcare.controllers.meals;
 
 import org.pesmypetcare.mypetcare.features.pets.Meals;
 import org.pesmypetcare.mypetcare.features.pets.Pet;
+import org.pesmypetcare.mypetcare.features.users.NotPetOwnerException;
 import org.pesmypetcare.mypetcare.features.users.User;
+import org.pesmypetcare.mypetcare.services.GoogleCalendarService;
 import org.pesmypetcare.mypetcare.services.MealManagerService;
+
+import java.util.concurrent.ExecutionException;
 
 public class TrDeleteMeal {
     private MealManagerService mealManagerService;
+    private GoogleCalendarService googleCalendarService;
     private User user;
     private Pet pet;
     private Meals meal;
 
-    public TrDeleteMeal(MealManagerService mealManagerService) {
+    public TrDeleteMeal(MealManagerService mealManagerService, GoogleCalendarService googleCalendarService) {
         this.mealManagerService = mealManagerService;
+        this.googleCalendarService = googleCalendarService;
     }
 
     /**
@@ -42,8 +48,12 @@ public class TrDeleteMeal {
     /**
      * Executes the transaction.
      */
-    public void execute() {
+    public void execute() throws NotPetOwnerException, ExecutionException, InterruptedException {
+        if (!user.getUsername().equals(pet.getOwner().getUsername())) {
+            throw new NotPetOwnerException();
+        }
         mealManagerService.deleteMeal(user, pet, meal);
         pet.deleteEvent(meal);
+        googleCalendarService.deleteEvent(pet, meal);
     }
 }
